fix(country-mode): split server-only getIntentCountry into its own file
Vercel build failed: lib/intent-context.ts imports next/headers (for
cookies() inside getIntentCountry), but Step 12 added a client-side
import of intentCountryFromSlug + quizKeyForIntentCode from the same
file in app/quiz/page.tsx. Turbopack rejects: "next/headers is only
available in Server Components".

Split:

- lib/intent-context.ts (shared): registry, type guard, slug/iso/quiz-
  key lookups, helpers. Pure — no next/headers, safe in client bundles.
- lib/intent-context-server.ts (new): the cookies()-using
  getIntentCountry function only. Server components import from here.

Updated 11 component imports to pull getIntentCountry from the new
location:
- components/country-mode/{CountryModeBanner,CountryListings/Experts/
  Compare/PopularLinks,CountryToolsStripWrapper}.tsx
- components/HomeCrossBorder.tsx
- components/HomePathfinder.tsx
- components/foreign-investment/{IntentCountryBadge,
  IntentCountryRecommendation}.tsx

Updated 5 test files to mock the new module (no importActual since the
new module has only one export):
- __tests__/components/{CountryHomepageWrappers,CountryModeBanner,
  CountryPopularLinks,HomeCrossBorder,HomePathfinder}.test.tsx
