+++ __tests__/components/CountryPopularLinks.test.tsx
@@ -6,15 +6,9 @@ const { mockGetIntentCountry, mockGetFilters } = vi.hoisted(() => ({
   mockGetFilters: vi.fn(),
 }));
 
-vi.mock("@/lib/intent-context", async () => {
-  const actual = await vi.importActual<typeof import("@/lib/intent-context")>(
-    "@/lib/intent-context",
-  );
-  return {
-    ...actual,
-    getIntentCountry: mockGetIntentCountry,
-  };
-});
+vi.mock("@/lib/intent-context-server", () => ({
+  getIntentCountry: mockGetIntentCountry,
+}));
 
 vi.mock("@/lib/country-mode", async () => {
   const actual = await vi.importActual<typeof import("@/lib/country-mode")>(
