+++ components/country-mode/CountryPopularLinks.tsx
@@ -14,10 +14,8 @@
  */
 
 import Link from "next/link";
-import {
-  getIntentCountry,
-  intentCountryMeta,
-} from "@/lib/intent-context";
+import { intentCountryMeta } from "@/lib/intent-context";
+import { getIntentCountry } from "@/lib/intent-context-server";
 import { getHomepageFiltersForCountry } from "@/lib/country-mode";
 
 export default async function CountryPopularLinks() {
